Rework logging
